+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ## Change log
 ----------------------
 
+Version 4.31.0
+-------------
+
+CHANGED:
+
+- handle interface case in method getTypeArguments
+- new getTypeArguments method with one argument created
+
 Version 4.30.0
 -------------